Move verify_openvm_stark into a new crate & refactor helper functions for verify_openvm_stark
