+++ Cargo.toml
@@ -66,6 +66,8 @@ members = [
     "extensions/pairing/transpiler",
     "extensions/pairing/guest",
     "extensions/pairing/tests",
+    "guest-libs/verify_stark/guest",
+    "guest-libs/verify_stark/tests",
 ]
 exclude = ["crates/sdk/example"]
 resolver = "2"
@@ -162,6 +164,7 @@ openvm-ecc-sw-macros = { path = "extensions/ecc/sw-macros", default-features = f
 openvm-pairing-circuit = { path = "extensions/pairing/circuit", default-features = false }
 openvm-pairing-transpiler = { path = "extensions/pairing/transpiler", default-features = false }
 openvm-pairing-guest = { path = "extensions/pairing/guest", default-features = false }
+openvm-verify-stark = { path = "guest-libs/verify_stark/guest", default-features = false }
 
 # Benchmarking
 openvm-benchmarks-utils = { path = "benchmarks/utils", default-features = false }
