+++ guest-libs/verify_stark/tests/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "openvm-verify-stark-integration-tests"
+description = "Integration tests for OpenVM verify-stark"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+
+[dev-dependencies]
+openvm-sdk = { workspace = true }
+openvm-circuit.workspace = true
+openvm-stark-sdk = { workspace = true }
+openvm-build.workspace = true
+openvm-native-compiler.workspace = true
+openvm-verify-stark.workspace = true
+eyre.workspace = true
+
+[features]
+default = ["openvm-circuit/parallel"]
