+++ guest-libs/verify_stark/tests/examples/verify_openvm_stark/Cargo.toml
@@ -5,9 +5,8 @@ version = "0.0.0"
 edition = "2021"
 
 [dependencies]
-openvm = { path = "../../../toolchain/openvm", features = ["std"] }
-hex-literal = { version = "0.4.1", default-features = false }
-bytemuck = { version = "1.20.0", features = ["extern_crate_alloc"] }
+openvm = { path = "../../../../../crates/toolchain/openvm", features = ["std"] }
+openvm-verify-stark = { path = "../../../guest" }
 
 [features]
 default = []
